Add spacing under header photo in About page
Introduced a divider with a height of 1.25px below the header photo on the About page to improve visual separation between the photo and the post title. This subtle spacing enhancement aids in creating a cleaner, more organized look, ensuring that the title stands out more clearly against the background of the header photo.

This adjustment aligns with design feedback aimed at improving content readability and aesthetics on the website.